LeetCode/src/johnny/algorithm/leetcode/Solution048.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package johnny.algorithm.leetcode;

/**
 * Rotate Image.
 * You are given an n x n 2D matrix representing an image.
 * 
 * Rotate the image by 90 degrees (clockwise).
 * 
 * Follow up:
 * Could you do this in-place?
 *  
 * @author Johnny
 */
public class Solution048 {
    public void rotate(int[][] matrix) {
        if (matrix == null || matrix.length == 0 || matrix[0].length == 0) {
            return;
        }
        
        //matrix[i][j] = matrix[n-1-j][i]
        int n = matrix.length;
        for (int i = 0; i < n / 2; i++) {
            for (int j = 0; j < Math.ceil(((double) n) / 2.); j++) {
                int temp = matrix[i][j];
                matrix[i][j] = matrix[n-1-j][i];
                matrix[n-1-j][i] = matrix[n-1-i][n-1-j];
                matrix[n-1-i][n-1-j] = matrix[j][n-1-i];
                matrix[j][n-1-i] = temp;
            }
        }        
    }
    public void rotate3(int[][] matrix) {
        if (matrix == null || matrix.length == 0 || matrix[0].length == 0) {
            return;
        }
        
        int n = matrix.length;
        int start = 0;
        int end = n - 1;
        while(start < end) {
            for(int i = start; i < end; i++) {
                int temp = matrix[start][i];
                matrix[start][i] = matrix[end - (i - start)][start];
                matrix[end - (i - start)][start] = matrix[end][end - (i - start)];
                matrix[end][end - (i - start)] = matrix[i][end];
                matrix[i][end] = temp;
            }
            start++;
            end--;
        }        
    }
     
    public void rotate2(int[][] matrix) {
        if (matrix == null || matrix.length == 0 || matrix[0].length == 0) {
            return;
        }
        
        int n = matrix.length;
        
        int[][] res = new int[n][n];
        for(int i = 0; i < n; i++) {
            for (int j = 0; j < n; j++) {
                res[j][n - i - 1] = matrix[i][j];
            }
        }
        
        //assign back
        for(int i = 0; i < n; i++){
            for(int j = 0; j < n; j++){
                matrix[i][j] = res[i][j];
            }
        } 
    }
}
